Chore: changed the class name to a more appropriate.
