+++ src/main/scala/io/snyk/plugin/datamodel/VulnerabilityCoordinate.scala
@@ -1,0 +1,36 @@
+package io.snyk.plugin.datamodel
+
+import io.circe.{Decoder, Encoder}
+
+case class VulnerabilityCoordinate(group: String, name: String, version: String) {
+  def unversionedName: String =  s"$group:$name"
+  def shortName: String =  s"$name@$version"
+  override def toString: String = s"$group:$name@$version"
+}
+
+object VulnerabilityCoordinate {
+  def from(fullVersionedName: String): VulnerabilityCoordinate = {
+    val versionArray = fullVersionedName.split('@')
+    val groupNameArray = versionArray.head.split(':')
+
+    val version = versionArray(1)
+    val group = groupNameArray(0)
+
+    if (groupNameArray.length > 1) {
+      val name = groupNameArray(1)
+
+      VulnerabilityCoordinate(group, name, version)
+    } else {
+      VulnerabilityCoordinate(group, "", version)
+    }
+  }
+
+  def from(fullName: String, version: String): VulnerabilityCoordinate = {
+    val arr = fullName.split(':')
+    VulnerabilityCoordinate(arr(0), arr(1), version)
+  }
+
+  implicit val encoder: Encoder[VulnerabilityCoordinate] = Encoder.encodeString.contramap[VulnerabilityCoordinate](_.toString)
+  implicit val decoder: Decoder[VulnerabilityCoordinate] = Decoder.decodeString.map(from)
+}
+
